fix GetStatsReport
