+++ Tests/Runtime/StatsReportTest.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 58cbe7a1d11171c43898ba79be53d6c9
+guid: 39f1c57a0918a08478d0c7e320283fa1
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
